changed file_id type to uint64_t

+++ src/server/services/optimizer/Optimizer.h
@@ -148,10 +148,10 @@ public:
 
     virtual void getTransferredBytes(std::map<Pair, double> &measureMap, time_t windowStart) = 0;
 
-    virtual void getOriginTransferredBytes(std::map<long long, double> &originInfo) = 0;
+    virtual void getOriginTransferredBytes(std::map<uint64_t, double> &originInfo) = 0;
 
     virtual void getTransferredBytesWithOrigin(
-        std::map<long long, double> &originInfo, 
+        std::map<uint64_t, double> &originInfo, 
         std::map<Pair, double> &measureMap,
         time_t windowStart
     ) = 0;
